refactor: 移動とエラーのバグ修正
