+++ main.go
@@ -5,7 +5,7 @@ import (
 	"github.com/jinzhu/gorm"
 	"github.com/joho/godotenv"
 	"go_training/config"
-	"go_training/lib/jwt_lib"
+	"go_training/infrastructure/jw_token"
 	"go_training/web/handler"
 	"log"
 )
@@ -20,7 +20,7 @@ func main() {
 
 func start() {
 	conf := config.NewConfig()
-	jwt_lib.KeyGenerator(conf)
+	jw_token.KeyGenerator(conf)
 	db, err := gorm.Open("mysql", conf.DB.GetSettingStr())
 	if err != nil {
 		panic(err.Error())
